+++ re_view/test/re_view/state_test.cljs
@@ -6,29 +6,57 @@
 
 (def append-el #(js/document.body.appendChild (js/document.createElement "div")))
 
-(deftest state-atom
-  (testing "React to local state atom"
+(deftest local-state
+  (let [el (append-el)
+        render #(v/render-to-dom % el)]
+    (testing "atom from initial-state"
+      (let [log (atom [])
+            local-state (atom nil)
+            view (v/view
+                   {:view/initial-state 0
+                    :view/did-mount     #(reset! local-state (:view/state %))}
+                   [{:keys [view/state]}]
+                   (swap! log conj @state)
+                   [:div "hello"])]
 
-    (let [log (atom [])
-          local-state (atom nil)
-          el (append-el)
-          view (v/view
-                 {:view/initial-state 0
-                  :view/did-mount     #(reset! local-state (:view/state %))}
-                 [{:keys [view/state]}]
-                 (swap! log conj @state)
-                 [:div "hello"])
-          render #(v/render-to-dom (view) el)]
+        (render (view))
+        (is (= @log [0]))
+        (swap! @local-state inc)
+        (v/flush!)
 
-      (render)
-      (is (= @log [0]))
-      (swap! @local-state inc)
-      (v/flush!)
+        (is (= @log [0 1]))
+        (reset! @local-state "x")
+        (v/flush!)
+        (is (= @log [0 1 "x"]))))
 
-      (is (= @log [0 1]))
-      (reset! @local-state "x")
-      (v/flush!)
-      (is (= @log [0 1 "x"])))))
+    (testing "passing an atom as state"
+      (let [the-atom (atom 111)
+            log (atom [])
+            view (v/view
+                   {:view/state the-atom}
+                   [this]
+                   (swap! log conj @(:view/state this))
+                   nil)]
+        (render (view))
+        (v/flush!)
+        (swap! the-atom inc)
+        (v/flush!)
+        (is (= @log [111 112]))))
+
+    (testing "passing an atom as :view/state prop"
+      (let [the-atom (atom 111)
+            log (atom [])
+            view (v/view
+                   [this]
+                   (swap! log conj @(:view/state this))
+                   nil)]
+        (render (view {:view/state the-atom}))
+        (v/flush!)
+        (swap! the-atom inc)
+        (v/flush!)
+        (is (= @log [111 112]))))))
+
+
 
 (deftest re-db